fs/ntfs3: Mark inode as bad as soon as error detected in mi_enum_attr()
[ Upstream commit 2afd4d267e6dbaec8d3ccd4f5396cb84bc67aa2e ]

Extended the `mi_enum_attr()` function interface with an additional
parameter, `struct ntfs_inode *ni`, to allow marking the inode
as bad as soon as an error is detected.

Upstream-Status: Pending

Reported-by: syzbot+73d8fc29ec7cba8286fa@syzkaller.appspotmail.com
Signed-off-by: Konstantin Komarov <almaz.alexandrovich@paragon-software.com>
[ The context change is introduced by the below commit:
a8f659ce14f8 ("fs/ntfs3: Add more checks in mi_enum_attr (part 2)")
which is irrelevant to the logic of this patch.
]
Signed-off-by: Qingfeng Hao <qingfeng.hao@windriver.com>
Signed-off-by: He Zhe <zhe.he@windriver.com>